requirements.txt
python-consul2
requests